requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
coreapi==2.3.1
coreschema==0.0.4
coverage==4.4.1
Django==1.11.1
django-braces==1.11.0
django-cors-middleware==1.3.1
django-crispy-forms==1.6.1
django-filter==1.0.4
django-guardian==1.4.8
django-nose==1.4.4
django-oauth-toolkit==0.12.0
djangorestframework==3.6.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
mysqlclient==1.3.10
nose==1.3.7
oauthlib==2.0.1
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
requests==2.16.5
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
